cloudrun_src/requirements.txt
uvicorn
fastapi
email-validator
google-cloud-resource-manager
google-cloud-tasks
google-cloud-billing

